Removed memory leaks cause by GameInterface, Saveable objects and Instancer not being deleted correctly Remove Component function should actually delete component pointer now Renderer is destroyed/initilaized only once instead of per scene

+++ engine/core/Timer.cpp
@@ -24,7 +24,7 @@ void Timer::UpdateTimer()
 unsigned int Timer::SleepTime(const unsigned int fps)
 {
 
-	const unsigned int milliSecsPerFrame = MILLI_TO_SEC / fps;
+	const unsigned int milliSecsPerFrame = static_cast<unsigned int>(MILLI_TO_SEC) / fps;
 
 	if (milliSecsPerFrame == 0)
 	{
